Support position based variants
This solves issue #902
